Add files via upload
A Dictionary application that will allow users to manage English words, their meaning and validate simple sentences. An initial dataset of words in a basic text file will be provided and will be expected to read the same into two (2) different appropriate and efficient data structures. The two data structures will be used throughout the application to benchmark and compare the efficiency of the operations performed on each.